+++ packages/eyes-sdk-core/lib/server/requestHelpers.js
@@ -24,6 +24,7 @@ const REQUEST_FAILED_CODES = ['ECONNRESET', 'ECONNABORTED', 'ETIMEDOUT', 'ENOTFO
 
 const CUSTOM_HEADER_NAMES = {
   REQUEST_ID: 'x-applitools-eyes-client-request-id',
+  AGENT_ID: 'x-applitools-eyes-client',
   EYES_EXPECT: 'Eyes-Expect',
   EYES_DATE: 'Eyes-Date',
 }
@@ -56,13 +57,23 @@ function configAxiosProxy({axiosConfig, proxy, logger}) {
 
   logger.log('proxy is set as http only, using tunnel', proxyObject.host, proxyObject.port)
 }
+
 function configAxiosFromConfiguration({axiosConfig, configuration, logger}) {
-  if (axiosConfig.params === undefined) {
-    axiosConfig.params = {}
-  }
-  if (axiosConfig.withApiKey && !('apiKey' in axiosConfig.params)) {
+  axiosConfig.params = axiosConfig.params || {}
+  axiosConfig.headers = axiosConfig.headers || {}
+
+  const {withApiKey, withAgentId} = axiosConfig.serverConfig
+  if (withApiKey) {
     axiosConfig.params.apiKey = configuration.getApiKey()
+  } else {
+    delete axiosConfig.params.apiKey
   }
+  if (withAgentId) {
+    axiosConfig.headers[CUSTOM_HEADER_NAMES.AGENT_ID] = configuration.getFullAgentId()
+  } else {
+    delete axiosConfig.headers[CUSTOM_HEADER_NAMES.AGENT_ID]
+  }
+
   if (!('removeSession' in axiosConfig.params)) {
     const removeSession = configuration.getRemoveSession()
     if (TypeUtils.isNotNull(removeSession)) {
@@ -82,10 +93,9 @@ function configAxiosFromConfiguration({axiosConfig, configuration, logger}) {
     }
   }
 }
-function configAxiosHeaders({axiosConfig}) {
-  if (axiosConfig.headers === undefined) {
-    axiosConfig.headers = {}
-  }
+
+function configureAxios({axiosConfig}) {
+  axiosConfig.headers = axiosConfig.headers || {}
   if (!(CUSTOM_HEADER_NAMES.REQUEST_ID in axiosConfig.headers)) {
     axiosConfig.headers[CUSTOM_HEADER_NAMES.REQUEST_ID] = axiosConfig.requestId
   }
@@ -204,7 +214,7 @@ async function handleRequestError({err, axios, logger}) {
 
 exports.configAxiosProxy = configAxiosProxy
 exports.configAxiosFromConfiguration = configAxiosFromConfiguration
-exports.configAxiosHeaders = configAxiosHeaders
+exports.configureAxios = configureAxios
 exports.delayRequest = delayRequest
 
 exports.handleRequestResponse = handleRequestResponse