+++ packages/eyes-selenium-3/src/Eyes.js
@@ -77,13 +77,9 @@
     EyesBase.call(this, this._promiseFactory, serverUrl || EyesBase.DEFAULT_EYES_SERVER, isDisabled)
   }
 
+  EyesBase.prototype._getBaseAgentId = () => 'eyes.selenium.javascript/' + VERSION
   Eyes.prototype = new EyesBase()
   Eyes.prototype.constructor = Eyes
-
-  //noinspection JSUnusedGlobalSymbols
-  Eyes.prototype._getBaseAgentId = function() {
-    return 'eyes.selenium.javascript/' + VERSION
-  }
 
   function _init(that, flow) {
     // Set PromiseFactory to work with the protractor control flow and promises