+++ packages/visual-grid-client/src/sdk/openEyes.js
@@ -171,6 +171,7 @@ function makeOpenEyes({
         apiKey,
         logHandler: logger.getLogHandler(),
         getBatchInfoWithCache,
+        agentId,
       })
 
     configureWrappers({
@@ -198,7 +199,6 @@ function makeOpenEyes({
       compareWithParentBranch,
       ignoreBaseline,
       serverUrl,
-      agentId,
     })
 
     if (!globalState.batchStore.hasCloseBatch()) {