+++ packages/visual-grid-client/src/sdk/wrapperUtils.js
@@ -2,10 +2,10 @@
 const EyesWrapper = require('./EyesWrapper')
 const {RectangleSize, TypeUtils} = require('@applitools/eyes-sdk-core')
 
-function initWrappers({count, apiKey, logHandler, getBatchInfoWithCache}) {
+function initWrappers({count, agentId, apiKey, logHandler, getBatchInfoWithCache}) {
   return Array.from(
     new Array(count),
-    () => new EyesWrapper({apiKey, logHandler, getBatchInfoWithCache}),
+    () => new EyesWrapper({agentId, apiKey, logHandler, getBatchInfoWithCache}),
   )
 }
 
@@ -51,7 +51,6 @@ function configureWrappers({
   compareWithParentBranch,
   ignoreBaseline,
   serverUrl,
-  agentId,
   useDom,
   enablePatterns,
   ignoreDisplacements,
@@ -87,7 +86,6 @@ function configureWrappers({
       wrapper.setCompareWithParentBranch(compareWithParentBranch)
     ignoreBaseline !== undefined && wrapper.setIgnoreBaseline(ignoreBaseline)
     serverUrl !== undefined && wrapper.setServerUrl(serverUrl)
-    agentId !== undefined && wrapper.setBaseAgentId(agentId)
   }
 }
 
@@ -126,10 +124,9 @@ function openWrappers({
 }
 
 function createRenderWrapper({serverUrl, apiKey, logHandler, proxy, agentId}) {
-  const wrapper = new EyesWrapper({apiKey, logHandler})
+  const wrapper = new EyesWrapper({agentId, apiKey, logHandler})
   serverUrl !== undefined && wrapper.setServerUrl(serverUrl)
   proxy !== undefined && wrapper.setProxy(proxy)
-  agentId !== undefined && wrapper.setBaseAgentId(agentId)
   return wrapper
 }